app/src/main/java/open/furaffinity/client/fragments/manageContactInfo.java: Uses a common class for dynamic fields app/src/main/java/open/furaffinity/client/fragments/manageUserPageAndProfileInformation.java: Uses a common class for dynamic fields app/src/main/java/open/furaffinity/client/utilities/dynamicEditItem.java: New shared class for dynamic fields app/src/main/java/open/furaffinity/client/utilities/uiControls.java: No real changes
