+++ app/src/main/java/open/furaffinity/client/fragments/manageUserPageAndProfileInformation.java
@@ -22,8 +22,10 @@ import java.util.concurrent.ExecutionException;
 
 import open.furaffinity.client.R;
 import open.furaffinity.client.pages.controlsProfile;
+import open.furaffinity.client.utilities.dynamicEditItem;
 import open.furaffinity.client.utilities.fabCircular;
 import open.furaffinity.client.utilities.kvPair;
+import open.furaffinity.client.utilities.uiControls;
 import open.furaffinity.client.utilities.webClient;
 
 public class manageUserPageAndProfileInformation extends Fragment {
@@ -38,82 +40,7 @@ public class manageUserPageAndProfileInformation extends Fragment {
 
     private int loadingStopCounter = 3;
 
-    private class editItem {
-        private LinearLayout linearLayout;
-        private TextView textView;
-        private EditText editText;
-        private Spinner spinner;
-        private String name;
-
-        private void initEditItem(Context context, LinearLayout linearLayout, String name, String header, String value, int maxLength, HashMap<String, String> options) {
-            this.name = name;
-
-            this.linearLayout = linearLayout;
-            this.textView = new TextView(context);
-            this.textView.setLayoutParams(new LinearLayout.LayoutParams(LinearLayout.LayoutParams.MATCH_PARENT, LinearLayout.LayoutParams.WRAP_CONTENT));
-            this.textView.setText(header);
-            this.linearLayout.addView(this.textView);
-
-            if(options == null) {
-                this.editText = new EditText(context);
-                this.editText.setText(value);
-
-                if (maxLength > 0) {
-                    InputFilter[] inputFilter = new InputFilter[1];
-                    inputFilter[0] = new InputFilter.LengthFilter(maxLength);
-                    this.editText.setFilters(inputFilter);
-                }
-
-                this.linearLayout.addView(this.editText);
-            } else  {
-                this.spinner = new Spinner(context);
-                open.furaffinity.client.utilities.uiControls.spinnerSetAdapter(context,this.spinner, options, value, false, false);
-                this.linearLayout.addView(this.spinner);
-            }
-        }
-
-        public editItem(Context context, LinearLayout linearLayout, String name, String header, String value) {
-            initEditItem(context, linearLayout, name, header, value, Integer.MAX_VALUE, null);
-        }
-
-        public editItem(Context context, LinearLayout linearLayout, String name, String header, String value, int maxLength) {
-            initEditItem(context, linearLayout, name, header, value, maxLength, null);
-        }
-
-        public editItem(Context context, LinearLayout linearLayout, String name, String header, HashMap<String, String> options) {
-            initEditItem(context, linearLayout, name, header, "", 0, options);
-        }
-
-        public editItem(Context context, LinearLayout linearLayout, String name, String header, String value, HashMap<String, String> options) {
-            initEditItem(context, linearLayout, name, header, value, 0, options);
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public String getValue(){
-            if(this.editText != null) {
-                return this.editText.getText().toString();
-            } else if (this.spinner != null) {
-                return ((kvPair)this.spinner.getSelectedItem()).getKey();
-            }
-
-            return "";
-        }
-
-        public void removeFromView() {
-            this.linearLayout.removeView(textView);
-
-            if(editText != null) {
-                this.linearLayout.removeView(this.editText);
-            } else if (spinner != null) {
-                this.linearLayout.removeView(this.spinner);
-            }
-        }
-    }
-
-    List<editItem> uiElementList;
+    List<dynamicEditItem> uiElementList;
 
     private void getElements(View rootView) {
         linearLayout = rootView.findViewById(R.id.linearLayout);
@@ -137,7 +64,7 @@ public class manageUserPageAndProfileInformation extends Fragment {
             }
 
             if (uiElementList != null) {
-                for (editItem currentItem : uiElementList) {
+                for (dynamicEditItem currentItem : uiElementList) {
                     currentItem.removeFromView();
                 }
             }
@@ -147,9 +74,9 @@ public class manageUserPageAndProfileInformation extends Fragment {
             if (page.getPageResults() != null) {
                 for (controlsProfile.inputItem currentInputItem : page.getPageResults()) {
                     if(currentInputItem.isSelect()) {
-                        uiElementList.add(new editItem(requireContext(), linearLayout, currentInputItem.getName(), currentInputItem.getHeader(), currentInputItem.getValue(), currentInputItem.getOptions()));
+                        uiElementList.add(new dynamicEditItem(requireContext(), linearLayout, currentInputItem.getName(), currentInputItem.getHeader(), currentInputItem.getValue(), currentInputItem.getOptions()));
                     } else {
-                        uiElementList.add(new editItem(requireContext(), linearLayout, currentInputItem.getName(), currentInputItem.getHeader(), currentInputItem.getValue(), currentInputItem.getMaxLength()));
+                        uiElementList.add(new dynamicEditItem(requireContext(), linearLayout, currentInputItem.getName(), currentInputItem.getHeader(), currentInputItem.getValue(), "", currentInputItem.getMaxLength()));
                     }
                 }
             }
@@ -164,7 +91,7 @@ public class manageUserPageAndProfileInformation extends Fragment {
                 params.put("do", "update");
                 params.put("key", page.getKey());
 
-                for(editItem currentItem : uiElementList) {
+                for(dynamicEditItem currentItem : uiElementList) {
                     params.put(currentItem.getName(), currentItem.getValue());
                 }
